gpt_engineer_enhanced_web/main/github_integration.py
from github import Github
from django.conf import settings

def connect_to_github(repo_name):
    g = Github(settings.GITHUB_ACCESS_TOKEN)
    user = g.get_user()
    repo = user.create_repo(repo_name)
    return repo.clone_url, repo.html_url

def push_to_github(repo_name, local_path):
    # Implement the logic to push local files to GitHub
    pass
